Marker for identifying call as follow-me loopback (#3977)
* Marker for identifying call as follow-me loopback

We need a way of the dialplan identifying calls that are generated by follow-me using loopback. This setting couple with a dialplan statement allows us to identify and set proper caller-id-name/number.

* Create 030_is_loopback.xml

Are you a follow_me call that is using the loopback interface? 
Ok then, are you a user on the system? 
Yes, ok. Let's set your outbound_caller_id_name and outbound_caller_id_number so that you show up correctly. 
Thanks.